pyproject.toml
[tool.poetry]
name = "payments"
version = "0.1.0"
description = "Website API"
authors = ["Mykola Fenyk <fenuk85@@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "0.29.0"
gunicorn = "^20.1.0"
fastapi = "0.110.3"
pydantic = "2.7.1"
psycopg2-binary = "2.9.9"
sqlalchemy = "2.0.29"
alembic = "1.13.1"
requests = "^2.28.0"
tenacity = "^6.1.0"
PyJWT = "2.8.0"
passlib = "1.7.4"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
aiofiles = "^23.1.0"
asyncpg = "^0.27.0"
httpx = "0.27.0"
python-multipart = "^0.0.6"
annotated-types = "0.6.0"
anyio = "4.3.0"
certifi = "2024.2.2"
click = "8.1.7"
exceptiongroup = "1.2.1"
greenlet = "3.0.3"
h11 = "0.14.0"
httpcore = "1.0.5"
idna = "3.7"
mako = "1.3.3"
markupsafe = "2.1.5"
pydantic-core = "2.18.2"
sniffio = "1.3.1"
starlette = "0.37.2"
typing-extensions = "4.11.0"
fastapi-pagination = "^0.12.24"


[tool.poetry.dev-dependencies]
mypy = "^0.971"
isort = "^5.12.0"
autoflake = "^1.5.3"
flake8 = "^5.0.4"
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.21.0"
asgi-lifespan = "^2.0.0"
devtools = "^0.11.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
balanced_wrapping = true
line_length = 90
profile = "black"

[tool.black]
line-length = 120
target-version = ['py38']

[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
ignore-magic = true
ignore-semiprivate = true
ignore-private = true
ignore-property-decorators = false
ignore-module = true
ignore-nested-functions = false
ignore-nested-classes = false
ignore-setters = false
fail-under = 100
exclude = [
    "/alembic",
    "/schema",
    "/models",
    "deps.py",
    "/tests"
]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*", "Settings.*"]
verbose = 2
quiet = false
whitelist-regex = []
omit-covered-files = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

